product filter fix

+++ backend/controllers/productController.js
@@ -18,9 +18,11 @@ exports.newProduct = catchAsyncErrors(async (req, res, next) => {
 
   let imagesLinks = [];
 
+  //**To solve the base 64 error: read this thread https://github.com/cloudinary/cloudinary_java/issues/156#issuecomment-525752814 */
   for (let i = 0; i < images.length; i++) {
-    const result = await cloudinary.v2.uploader.upload(images[i], {
+    const result = await cloudinary.v2.uploader.upload_large(images[i], {
       folder: 'products',
+      chunk_size: 6000000,
     });
 
     imagesLinks.push({
@@ -41,7 +43,7 @@ exports.newProduct = catchAsyncErrors(async (req, res, next) => {
 });
 
 exports.getProducts = async (req, res, next) => {
-  const resPerPage = 8;
+  const resPerPage = 4;
   const productCount = await Product.countDocuments();
   const apiFeatures = new APIFeatures(Product.find(), req.query)
     .search()
@@ -88,37 +90,94 @@ exports.getSingleProduct = async (req, res, next) => {
 
 ///update product => api/v1/product:id
 
-exports.UpdateProduct = async (req, res, next) => {
+exports.UpdateProduct = catchAsyncErrors(async (req, res, next) => {
   let product = await Product.findById(req.params.id);
+
   if (!product) {
     return next(new ErrorHandler('Product not found', 404));
   }
+
+  let images = [];
+  if (typeof req.body.images === 'string') {
+    images.push(req.body.images);
+  } else {
+    images = req.body.images;
+  }
+
+  if (images !== undefined) {
+    // Deleting images associated with the product
+    for (let i = 0; i < product.images.length; i++) {
+      const result = await cloudinary.v2.uploader.destroy(
+        product.images[i].public_id
+      );
+    }
+
+    let imagesLinks = [];
+
+    for (let i = 0; i < images.length; i++) {
+      const result = await cloudinary.v2.uploader.upload(images[i], {
+        folder: 'products',
+      });
+
+      imagesLinks.push({
+        public_id: result.public_id,
+        url: result.secure_url,
+      });
+    }
+
+    req.body.images = imagesLinks;
+  }
+
   product = await Product.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
-    runValidatiors: false,
+    runValidators: true,
     useFindAndModify: false,
   });
+
   res.status(200).json({
     success: true,
     product,
   });
-};
+});
+
+// Delete Product   =>   /api/v1/admin/product/:id
+exports.deleteProduct = catchAsyncErrors(async (req, res, next) => {
+  const product = await Product.findById(req.params.id);
+
+  if (!product) {
+    return next(new ErrorHandler('Product not found', 404));
+  }
+
+  // Deleting images associated with the product
+  for (let i = 0; i < product.images.length; i++) {
+    const result = await cloudinary.v2.uploader.destroy(
+      product.images[i].public_id
+    );
+  }
+
+  await product.remove();
+
+  res.status(200).json({
+    success: true,
+    message: 'Product is deleted.',
+  });
+});
 
 //Delete product => /api/v1/admin/product/id
 
 exports.deleteProduct = catchAsyncErrors(async (req, res, next) => {
   const product = await Product.findById(req.params.id);
 
-  // if (!product) {
-  //   return next(new ErrorHandler('Product not found', 404));
-  // }
+  if (!product) {
+    return next(new ErrorHandler('Product not found', 404));
+  }
 
-  // // Deleting images associated with the product
-  // for (let i = 0; i < product.images.length; i++) {
-  //   const result = await cloudinary.v2.uploader.destroy(
-  //     product.images[i].public_id
-  //   );
-  // }
+  // Deleting images associated with the product
+  for (let i = 0; i < product.images.length; i++) {
+    const result = await cloudinary.v2.uploader.destroy(
+      product.images[i].public_id
+    );
+  }
 
   await product.deleteOne();
